Use a vacuum process to clean up tiddlers from the index
When a tiddler is deleted, it's not removed from the lunr index,
which can be annoying since it will still show up in search results.
The old solution to this was to filter out search hits that were no
longer in the store, but this hurt search result rendering performance
pretty badly.  Ideally, we would remove the tiddler from the index in
response to a deletion hook - unfortunately, it doesn't seem TiddlyWiki
has such a hook yet.  Until such time, I've added a "background job"
that will clean up removed tiddlers.