Patch lunr to fix bug
There's a bug in lunr 1.0.0 that that causes lunr.Index.load to
improperly load a chunk of data generated via lunr.Index.toJSON - in
short, lunr.SortedSet.toJSON isn't called when it should be, but
lunr.SortedSet.load *is*, and the JSON representation it decides on
is incompatible with its internal format, which results in the resulting
objects having the wrong shape after a load.  lunr 2.0.x is out, but it
changes the API considerably - I'll deliberate on whether to upgrade to
this at a later time.