Fixes for TypeScript 2.5+
Between TypeScript 2.4 and 2.5 (in commit
https://github.com/Microsoft/TypeScript/commit/680bfbb705, to be exact),
TypeScript adjusted its definition of whether a module name was relative
or not to include "rooted disk paths".
$:/core/modules/widgets/widget.js" looks like a rooted disk path, so the
compiler treats it as a relative module path and throws TS2436.

So, in an effort to make my code compile under more recent versions of
TypeScript, I'm just going to lift my ambient module definition out so
that I can use it directly in fts-action-generate-index.ts