Don't mix object/array behavior in expansion tree
Before this change, I was relying on the object-like nature of
JavaScript arrays for my query expansion trees - the object keys
of a node in the tree were the next links in the expansion chain for
multi-word expansions, and the array indices were the terms that that
point in the expansion chain would expand to.  This worked great...until
my code encountered expansions using numerical tokens.  That case broke
lunr.js, because I guess JS coerced "2" to the array index 2 or
something and filled them with nulls, and lunr.js didn't like that.
Serves me right for playing with arrays and objects in such an ambiguous
way!

Anyway, the new code just stored the resulting token expansion in a key
named ".expansion" - I don't *think* this should pop up naturally,
especially since I think the lunr.js trimmer will trim the leading dot