Patch lunr to duck type arrays
Meaning check if an object has the 'length' property rather than if it's
"instanceof Array".  This is used in lunr's token pipeline, which allows
you to insert arbitrary functions.  The expandQuery function is
generated by a separate TiddlyWiki module, which is loaded using a
separate VM when TiddlyWiki is running under Node, so "o instanceof
Array" will always evaluate to false, even though o is actually an
Array!